+++ compiler/src/Language/Bond/Codegen/Templates.hs
@@ -45,6 +45,7 @@ module Language.Bond.Codegen.Templates
     , comm_h
     , comm_cpp
     , grpc_h
+    , grpc_cpp
       -- ** C#
     , FieldMapping(..)
     , StructMapping(..)
@@ -65,6 +66,7 @@ import Language.Bond.Codegen.Cpp.Types_cpp
 import Language.Bond.Codegen.Cpp.Types_h
 import Language.Bond.Codegen.Cpp.Comm_cpp
 import Language.Bond.Codegen.Cpp.Comm_h
+import Language.Bond.Codegen.Cpp.Grpc_cpp
 import Language.Bond.Codegen.Cpp.Grpc_h
 import Language.Bond.Codegen.Cs.Types_cs
 import Language.Bond.Codegen.Cs.Comm_cs